backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.4
pytest-anyio==0.0.0
anyio==4.7.0
black==24.10.0
ruff==0.8.4
